Default Lucid to 2.0.5, and use Rails Machine mirror
Previously, we used 10gen's package repository desribed at
http://www.mongodb.org/display/DOCS/Ubuntu+and+Debian+packages

The main problem with it is that it only contains the _latest_ release.

moonshine_mongodb was defaulting to installing the latest, which will
make sure that new servers always get a working package version. The
downside is that it would cause a surprise upgrade on deploy if a
release has come out.

This commit locks to 2.0.5, and now uses a mirror we created containing
just 2.0.5

Signed-off-by: Rails Machine <ops@railsmachine.com>